user_interface/web-server/requirements.txt
jsonlines
pyyaml
pydash
paho-mqtt
quart
quart-cors
quart-openapi
